Changed interrupts, display, etc.

+++ StoicGoose/Emulation/CPU/V30MZ.ModRM.cs
@@ -45,7 +45,7 @@
 
 				case ModRM.Modes.TwoByteDisplacement:
 					{
-						var displacement = ReadMemory16(cs, ip);
+						var displacement = (short)ReadMemory16(cs, ip);
 						ip += 2;
 						switch (modRm.Mem)
 						{
